ADDED: Interface for feeds
Eventually, I would like to have all of the different types of feeds
implement this interface so that I can depend on methods in the client
regardless of which feed it comes from.  For example getTitle() should
always return the title of the feed whether it's from the FeedService,
GoogleReader, or iTunes.